Remove old correlation code
Fix access of threshold and slope in evaluation forecast

None implement the desired the type of correlation. The type of
correlation we were hoping to induce was some form of deterministic
success pattern for tasks, but along two dimensions, one as measured
and a second latent dimension we don't see but is assumed to exist.